Testing an enum < 0 (such as vis in hTvFromString()) doesn't work. Instead, cast the enum to an int when testing it.

+++ src/hg/lib/hui.c
@@ -92,7 +92,7 @@ enum trackVisibility hTvFromString(char *s)
  * equivalent enum. */
 {
 enum trackVisibility vis = hTvFromStringNoAbort(s);
-if (vis < 0)
+if ((int)vis < 0)
    errAbort("Unknown visibility %s", s);
 return vis;
 }
